# docs/swagger.yaml
basePath: /
definitions:
  handlers.createOfficerRequest:
    properties:
      image_uri:
        type: string
      linkedin_photo:
        type: string
      name:
        type: string
      title:
        type: string
    required:
    - name
    - title
    type: object
  handlers.errorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.getOfficersResponse:
    properties:
      officers:
        items:
          $ref: '#/definitions/handlers.officerResponse'
        type: array
    type: object
  handlers.officerResponse:
    properties:
      id:
        type: integer
      image_uri:
        type: string
      linkedin_photo:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
  description: CMAC's Go API
  title: Service Main API
  version: "1.0"
paths:
  /auth/officers:
    post:
      consumes:
      - application/json
      description: Create a new officer record.
      parameters:
      - description: Officer payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.createOfficerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.officerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - SessionAuth: []
      summary: Create officer
      tags:
      - officers
  /officers:
    get:
      description: Returns a list of every officer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.getOfficersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: List all officers
      tags:
      - officers
securityDefinitions:
  SessionAuth:
    in: cookie
    name: sessid
    type: apiKey
swagger: "2.0"